+++ packages/components/src/Combobox/ComboboxProvider.tsx
@@ -1,11 +1,14 @@
 import React, { RefObject, useRef, useState } from "react";
 import styles from "./Combobox.css";
+import { ComboboxOption } from "./Combobox.types";
 
 export const ComboboxContext = React.createContext(
   {} as {
     multiselect: boolean;
     open: boolean;
     setOpen: (open: boolean) => void;
+    selected: ComboboxOption[];
+    setSelected: (open: ComboboxOption[]) => void;
     wrapperRef: RefObject<HTMLDivElement>;
   },
 );
@@ -20,11 +23,12 @@ export function ComboboxContextProvider(
 ): JSX.Element {
   const multiselect = props.multiselect || false;
   const [open, setOpen] = useState<boolean>(false);
+  const [selected, setSelected] = useState<ComboboxOption[]>([]);
   const wrapperRef = useRef<HTMLDivElement>(null);
 
   return (
     <ComboboxContext.Provider
-      value={{ multiselect, open, setOpen, wrapperRef }}
+      value={{ multiselect, open, setOpen, wrapperRef, selected, setSelected }}
     >
       <div ref={wrapperRef}>
         {open && (